Hides add song button as well
