+++ src/modules/common/zindex.js
@@ -2,8 +2,8 @@ const baseZIndexList = [
   0,
   'toast',
   'player-expanded',
-  'controls',
   'iframeblocker',
+  'controls',
   'player',
   'header',
   'nav',
